Better file modularization
